// src/app/views/personal/info/info.ts
import { Component, inject } from '@angular/core';
import { FormBuilder, Validators } from '@angular/forms';
import { PaymentType } from '../../../../types/payment.type';
import { DeliveryType } from '../../../../types/delivery.type';
import { UserService } from '../../../shared/services/user-service';
import { DefaultResponseType } from '../../../../types/default-response.type';
import { UserInfoType } from '../../../../types/user-info.type';
import { MatSnackBar } from '@angular/material/snack-bar';
import { HttpErrorResponse } from '@angular/common/http';

@Component({
  selector: 'app-info',
  standalone: false,
  templateUrl: './info.html',
  styleUrl: './info.scss',
})
export class InfoComponent {
  private readonly fb = inject(FormBuilder);
  private readonly userService = inject(UserService);
  private readonly _snackBar = inject(MatSnackBar);

  protected deliveryType: DeliveryType = DeliveryType.delivery;
  protected deliveryTypes = DeliveryType;
  protected paymentTypes = PaymentType;
  protected userInfoForm = this.fb.group({
    firstName: [''],
    lastName: [''],
    fatherName: [''],
    phone: [''],
    paymentType: [PaymentType.cashToCourier],
    email: ['', [Validators.required]],
    street: [''],
    house: [''],
    entrance: [''],
    apartment: [''],
  });

  constructor() {
    this.userService.getUserInfo()
      .subscribe({
        next: (data: UserInfoType | DefaultResponseType) => {
          if ((data as DefaultResponseType).error !== undefined) throw new Error((data as DefaultResponseType).message);
          const userInfo = data as UserInfoType;
          const paramsToUpdate = {
            firstName: userInfo.firstName ? userInfo.firstName : '',
            lastName: userInfo.lastName ? userInfo.lastName : '',
            fatherName: userInfo.fatherName ? userInfo.fatherName : '',
            phone: userInfo.phone ? userInfo.phone : '',
            paymentType: userInfo.paymentType ? userInfo.paymentType : PaymentType.cashToCourier,
            email: userInfo.email ? userInfo.email : '',
            street: userInfo.street ? userInfo.street : '',
            house: userInfo.house ? userInfo.house : '',
            entrance: userInfo.entrance ? userInfo.entrance : '',
            apartment: userInfo.apartment ? userInfo.apartment : '',
          };
          this.userInfoForm.setValue(paramsToUpdate);
          if (userInfo.deliveryType) this.deliveryType = userInfo.deliveryType;
        }
      })
  }

  protected changeDeliveryType(type: DeliveryType) {
    this.deliveryType = type;
    this.userInfoForm.markAsDirty();
  }

  protected updateUserInfo() {
    if (this.userInfoForm.valid) {
      const paramObject: UserInfoType = {
        email: this.userInfoForm.value.email ? this.userInfoForm.value.email : '',
        deliveryType: this.deliveryType,
        paymentType: this.userInfoForm.value.paymentType ? this.userInfoForm.value.paymentType : this.paymentTypes.cashToCourier
      }
      if (this.userInfoForm.value.firstName) paramObject.firstName = this.userInfoForm.value.firstName;
      if (this.userInfoForm.value.lastName) paramObject.lastName = this.userInfoForm.value.lastName;
      if (this.userInfoForm.value.fatherName) paramObject.fatherName = this.userInfoForm.value.fatherName;
      if (this.userInfoForm.value.phone) paramObject.phone = this.userInfoForm.value.phone;
      if (this.userInfoForm.value.street) paramObject.street = this.userInfoForm.value.street;
      if (this.userInfoForm.value.house) paramObject.house = this.userInfoForm.value.house;
      if (this.userInfoForm.value.entrance) paramObject.entrance = this.userInfoForm.value.entrance;
      if (this.userInfoForm.value.apartment) paramObject.apartment = this.userInfoForm.value.apartment;
      this.userService.updateUserInfo(paramObject)
        .subscribe({
          next: (data: DefaultResponseType) => {
            if (data.error) {
              this._snackBar.open(data.message);
              throw new Error(data.message);
            }
            this._snackBar.open('Данные успешно сохранены');
            this.userInfoForm.markAsPristine();
          },
          error: (errorResponse: HttpErrorResponse) => {
            if (errorResponse.error && errorResponse.error.message) {
              this._snackBar.open(errorResponse.error.message);
            } else {
              this._snackBar.open('Ошибка сохранения');
            }
          }
        });

    }
  }
}
